+++ CHANGELOG.md
@@ -1,3 +1,13 @@
 ## [Unreleased]
 
-- Lay down the groundwork for a new gem...
+## [0.1.0] Initial Release - 2023-05-20
+
+- Define basic functionality for this gem:
+  - `Say.say(message, type = nil, &block)`
+  - `Say.say_with_block(header_message, footer_message: "Done")`
+  - `Say.say_header(message = nil, **kwargs)`
+  - `Say.say_item(message, **kwargs)`
+  - `Say.say_footer(message = "Done", **kwargs)`
+  - `Say.build_banner(message = nil, columns: MAX_COLUMNS)`
+  - `Say.build_message(message, type: nil)`
+  - `Say.do_say(*messages, silent: false)`